+++ build-config/fragments/htmlinit.js
@@ -1,20 +1,18 @@
 /* eslint-env node */
 
 const path = require('path');
+const banner = require('./banner');
+
+const rootDir = path.resolve(__dirname, '..', '..');
 
 module.exports = {
     entry: {
-        'html-init': path.resolve(
-            __dirname,
-            '../',
-            '../',
-            'src',
-            'html-init.js'
-        )
+        'html-init': path.join(rootDir, 'src', 'html-init.js')
     },
     output: {
-        path: path.resolve(__dirname, '../', '../', 'dist'),
+        path: path.join(rootDir, 'dist'),
         filename: 'wavesurfer-[name].js',
         library: ['WaveSurfer', '[name]']
-    }
+    },
+    plugins: [banner.libBanner]
 };