+++ src/plugin/cursor.js
@@ -10,6 +10,9 @@
  * @property {number} zIndex=3 The z-index of the cursor element
  * @property {object} customStyle An object with custom styles which are applied
  * to the cursor element
+ * @property {boolean} showTime=false Show the time on the cursor.
+ * @property {object} customShowTimeStyle An object with custom styles which are
+ * applied to the cursor time element.
  */
 
 /**
@@ -73,7 +76,9 @@ export default class CursorPlugin {
         opacity: '0.25',
         style: 'solid',
         zIndex: 4,
-        customStyle: {}
+        customStyle: {},
+        customShowTimeStyle: {},
+        showTime: false
     };
 
     /** @private */
@@ -99,11 +104,23 @@ export default class CursorPlugin {
         /** @private */
         this.style = ws.util.style;
         /**
-         * The cursor html element
+         * The cursor HTML element
          *
          * @type {?HTMLElement}
          */
         this.cursor = null;
+        /**
+         * displays the time next to the cursor
+         *
+         * @type {Boolean}
+         */
+        this.showTime = null;
+        /**
+         * The html container that will display the time
+         *
+         * @type {?HTMLElement}
+         */
+        this.displayTime = null;
         /** @private */
         this.params = ws.util.extend({}, this.defaultParams, params);
     }
@@ -112,8 +129,8 @@ export default class CursorPlugin {
      * Initialise the plugin (used by the Plugin API)
      */
     init() {
-        const wrapper = this.wavesurfer.container;
-        this.cursor = wrapper.appendChild(
+        this.wrapper = this.wavesurfer.container;
+        this.cursor = this.wrapper.appendChild(
             this.style(
                 document.createElement('cursor'),
                 this.wavesurfer.util.extend(
@@ -124,7 +141,7 @@ export default class CursorPlugin {
                         top: 0,
                         bottom: 0,
                         width: '0',
-                        display: 'block',
+                        display: 'flex',
                         borderRightStyle: this.params.style,
                         borderRightWidth: this.params.width,
                         borderRightColor: this.params.color,
@@ -135,11 +152,46 @@ export default class CursorPlugin {
                 )
             )
         );
+        if (this.params.showTime) {
+            this.showTime = this.wrapper.appendChild(
+                this.style(
+                    document.createElement('showTitle'),
+                    this.wavesurfer.util.extend(
+                        {
+                            position: 'absolute',
+                            zIndex: this.params.zIndex,
+                            left: 0,
+                            top: 0,
+                            bottom: 0,
+                            width: 'auto',
+                            display: 'flex',
+                            opacity: this.params.opacity,
+                            pointerEvents: 'none',
+                            height: '100%'
+                        },
+                        this.params.customStyle
+                    )
+                )
+            );
+            this.displayTime = this.showTime.appendChild(
+                this.style(
+                    document.createElement('div'),
+                    this.wavesurfer.util.extend(
+                        {
+                            display: 'inline',
+                            pointerEvents: 'none',
+                            margin: 'auto'
+                        },
+                        this.params.customShowTimeStyle
+                    )
+                )
+            );
+        }
 
-        wrapper.addEventListener('mousemove', this._onMousemove);
+        this.wrapper.addEventListener('mousemove', this._onMousemove);
         if (this.params.hideOnBlur) {
-            wrapper.addEventListener('mouseenter', this._onMouseenter);
-            wrapper.addEventListener('mouseleave', this._onMouseleave);
+            this.wrapper.addEventListener('mouseenter', this._onMouseenter);
+            this.wrapper.addEventListener('mouseleave', this._onMouseleave);
         }
     }
 
@@ -147,6 +199,9 @@ export default class CursorPlugin {
      * Destroy the plugin (used by the Plugin API)
      */
     destroy() {
+        if (this.params.showTime) {
+            this.cursor.parentNode.removeChild(this.showTime);
+        }
         this.cursor.parentNode.removeChild(this.cursor);
         this.wrapper.removeEventListener('mousemove', this._onMousemove);
         if (this.params.hideOnBlur) {
@@ -164,6 +219,18 @@ export default class CursorPlugin {
         this.style(this.cursor, {
             left: `${pos}px`
         });
+        if (this.params.showTime) {
+            const duration = this.wavesurfer.getDuration();
+            const elementWidth =
+                this.wavesurfer.drawer.getWidth() /
+                this.wavesurfer.params.pixelRatio;
+            const timeValue = Math.max(0, (pos / elementWidth) * duration);
+            const formatValue = this.formatTime(timeValue);
+            this.style(this.showTime, {
+                left: `${pos}px`
+            });
+            this.displayTime.innerHTML = `${formatValue}`;
+        }
     }
 
     /**
@@ -171,8 +238,13 @@ export default class CursorPlugin {
      */
     showCursor() {
         this.style(this.cursor, {
-            display: 'block'
+            display: 'flex'
         });
+        if (this.params.showTime) {
+            this.style(this.showTime, {
+                display: 'flex'
+            });
+        }
     }
 
     /**
@@ -182,5 +254,21 @@ export default class CursorPlugin {
         this.style(this.cursor, {
             display: 'none'
         });
+        if (this.params.showTime) {
+            this.style(this.showTime, {
+                display: 'none'
+            });
+        }
+    }
+
+    /** timestamp time calculation */
+    formatTime(cursorTime) {
+        return [cursorTime].map(time =>
+            [
+                Math.floor((time % 3600) / 60), // minutes
+                ('00' + Math.floor(time % 60)).slice(-2), // seconds
+                ('000' + Math.floor((time % 1) * 1000)).slice(-3) // miliseconds
+            ].join(':')
+        );
     }
 }