+++ src/plugin/spectrogram.js
@@ -5,7 +5,7 @@
 const FFT = function(bufferSize, sampleRate, windowFunc, alpha) {
     this.bufferSize = bufferSize;
     this.sampleRate = sampleRate;
-    this.bandwidth = 2 / bufferSize * sampleRate / 2;
+    this.bandwidth = (2 / bufferSize) * (sampleRate / 2);
 
     this.sinTable = new Float32Array(bufferSize);
     this.cosTable = new Float32Array(bufferSize);
@@ -19,8 +19,7 @@ const FFT = function(bufferSize, sampleRate, windowFunc, alpha) {
         case 'bartlett':
             for (var i = 0; i < bufferSize; i++) {
                 this.windowValues[i] =
-                    2 /
-                    (bufferSize - 1) *
+                    (2 / (bufferSize - 1)) *
                     ((bufferSize - 1) / 2 - Math.abs(i - (bufferSize - 1) / 2));
             }
             break;
@@ -29,7 +28,7 @@ const FFT = function(bufferSize, sampleRate, windowFunc, alpha) {
                 this.windowValues[i] =
                     0.62 -
                     0.48 * Math.abs(i / (bufferSize - 1) - 0.5) -
-                    0.38 * Math.cos(Math.PI * 2 * i / (bufferSize - 1));
+                    0.38 * Math.cos((Math.PI * 2 * i) / (bufferSize - 1));
             }
             break;
         case 'blackman':
@@ -37,14 +36,15 @@ const FFT = function(bufferSize, sampleRate, windowFunc, alpha) {
             for (var i = 0; i < bufferSize; i++) {
                 this.windowValues[i] =
                     (1 - alpha) / 2 -
-                    0.5 * Math.cos(Math.PI * 2 * i / (bufferSize - 1)) +
-                    alpha / 2 * Math.cos(4 * Math.PI * i / (bufferSize - 1));
+                    0.5 * Math.cos((Math.PI * 2 * i) / (bufferSize - 1)) +
+                    (alpha / 2) *
+                        Math.cos((4 * Math.PI * i) / (bufferSize - 1));
             }
             break;
         case 'cosine':
             for (var i = 0; i < bufferSize; i++) {
                 this.windowValues[i] = Math.cos(
-                    Math.PI * i / (bufferSize - 1) - Math.PI / 2
+                    (Math.PI * i) / (bufferSize - 1) - Math.PI / 2
                 );
             }
             break;
@@ -56,7 +56,7 @@ const FFT = function(bufferSize, sampleRate, windowFunc, alpha) {
                     -0.5 *
                         Math.pow(
                             (i - (bufferSize - 1) / 2) /
-                                (alpha * (bufferSize - 1) / 2),
+                                ((alpha * (bufferSize - 1)) / 2),
                             2
                         )
                 );
@@ -65,21 +65,22 @@ const FFT = function(bufferSize, sampleRate, windowFunc, alpha) {
         case 'hamming':
             for (var i = 0; i < bufferSize; i++) {
                 this.windowValues[i] =
-                    0.54 - 0.46 * Math.cos(Math.PI * 2 * i / (bufferSize - 1));
+                    (0.54 - 0.46) *
+                    Math.cos((Math.PI * 2 * i) / (bufferSize - 1));
             }
             break;
         case 'hann':
         case undefined:
             for (var i = 0; i < bufferSize; i++) {
                 this.windowValues[i] =
-                    0.5 * (1 - Math.cos(Math.PI * 2 * i / (bufferSize - 1)));
+                    0.5 * (1 - Math.cos((Math.PI * 2 * i) / (bufferSize - 1)));
             }
             break;
         case 'lanczoz':
             for (var i = 0; i < bufferSize; i++) {
                 this.windowValues[i] =
-                    Math.sin(Math.PI * (2 * i / (bufferSize - 1) - 1)) /
-                    (Math.PI * (2 * i / (bufferSize - 1) - 1));
+                    Math.sin(Math.PI * ((2 * i) / (bufferSize - 1) - 1)) /
+                    (Math.PI * ((2 * i) / (bufferSize - 1) - 1));
             }
             break;
         case 'rectangular':
@@ -90,8 +91,7 @@ const FFT = function(bufferSize, sampleRate, windowFunc, alpha) {
         case 'triangular':
             for (var i = 0; i < bufferSize; i++) {
                 this.windowValues[i] =
-                    2 /
-                    bufferSize *
+                    (2 / bufferSize) *
                     (bufferSize / 2 - Math.abs(i - (bufferSize - 1) / 2));
             }
             break;
@@ -221,7 +221,7 @@ const FFT = function(bufferSize, sampleRate, windowFunc, alpha) {
  * @property {string|HTMLElement} container Selector of element or element in
  * which to render
  * @property {number} fftSamples=512 number of samples to fetch to FFT. Must be
- * a pwer of 2.
+ * a power of 2.
  * @property {number} noverlap Size of the overlapping window. Must be <
  * fftSamples. Auto deduced from canvas size by default.
  * @property {string} windowFunc='hann' The window function to be used. One of
@@ -569,7 +569,7 @@ export default class SpectrogramPlugin {
 
             const freq = freqStart + step * i;
             const index = Math.round(
-                freq / (this.sampleRate / 2) * this.fftSamples
+                (freq / (this.sampleRate / 2)) * this.fftSamples
             );
             const label = this.freqType(freq);
             const units = this.unitType(freq);
@@ -636,7 +636,7 @@ export default class SpectrogramPlugin {
                         if (column[k] == null) {
                             column[k] = 0;
                         }
-                        column[k] += overlap / newPiece * oldMatrix[j][k];
+                        column[k] += (overlap / newPiece) * oldMatrix[j][k];
                     }
                 }
                 /* eslint-enable max-depth */