+++ src/util/index.js
@@ -8,3 +8,4 @@ export { default as style } from './style';
 export { default as requestAnimationFrame } from './request-animation-frame';
 export { default as frame } from './frame';
 export { default as debounce } from 'debounce';
+export { default as preventClick } from './prevent-click';